Misc PR adjustments; extract StockValidatioError to custom /errors folder + addtl. cleanup.
