+++ src/errors/StockValidationError.ts
@@ -1,0 +1,6 @@
+export class StockValidationError extends Error {
+  constructor(message: string, public itemId?: string) {
+    super(message);
+    this.name = 'StockValidationError';
+  }
+}
