+++ src/helpers/order.ts
@@ -1,11 +1,5 @@
 import { StockLevelType } from "../models/enums/stockLevelType";
-
-export class StockValidationError extends Error {
-  constructor(message: string, public itemId?: string) {
-    super(message);
-    this.name = 'StockValidationError';
-  }
-}
+import { StockValidationError } from "../errors/StockValidationError";
 
 export function getUpdatedStockLevel(
   currentLevel: StockLevelType,
